Fix llvm to spirv translation for image types
Image type may be represented in llvm IR in two forms:
- OpenCL form, for example, '%opencl.image2d_ro_t',
- SPIRV form, for example, '%spirv.Image._void_1_0_0_0_0_0_0',
but in both cases it is the same image type and it should be mapped to
one SPIR-V type.

The patch prevents creation two SPIR-V types for one LLVM represented in
two forms checking both forms in the types map.